backend_python/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.45
PyJWT==2.8.0
bcrypt==4.0.1
python-dotenv==1.0.0

